gpt/services/chat/requirements.txt
groq
python-dotenv
fastapi
jinja2
python-multipart
uvicorn